fix(cloud): use correct property to retrive datacenter endpoint
